hello_world/requirements.txt
aws-psycopg2 == 1.3.8
